Initial support for AVRO consumer
