+++ pom.xml
@@ -10,7 +10,7 @@
     <packaging>jar</packaging>
 
     <properties>
-        <revision>0.1.2</revision>
+        <revision>0.1.3</revision>
         <java.version>1.8</java.version>
         <project.build.sourceEncoding>UTF-8</project.build.sourceEncoding>
         <kafka-clients.version>2.4.0</kafka-clients.version>
@@ -28,6 +28,13 @@
             <url>file://${project.build.directory}/mvn-repo</url>
         </repository>
     </distributionManagement>
+
+    <repositories>
+        <repository>
+            <id>confluent</id>
+            <url>http://packages.confluent.io/maven/</url>
+        </repository>
+    </repositories>
 
     <dependencies>
         <!-- https://mvnrepository.com/artifact/org.apache.kafka/kafka-clients -->
@@ -72,6 +79,14 @@
             <groupId>com.jayway.jsonpath</groupId>
             <artifactId>json-path</artifactId>
             <version>2.4.0</version>
+        </dependency>
+
+        <dependency>
+            <groupId>io.confluent</groupId>
+            <artifactId>kafka-avro-serializer</artifactId>
+            <version>5.5.0</version>
+            <!-- Avoid including it for everybody -->
+            <optional>true</optional>
         </dependency>
 
     </dependencies>
